atualização atividade 19

+++ 19/controller/PedidoController.php
@@ -1,15 +1,14 @@
 <?php
-
-require_once 'database/PedidoRepository.php'; 
+require_once 'database/PedidoRepository.php';
 
 class PedidoController {
-    public static function handleRequest($action){
-        switch ($action){
+    public static function handleRequest($action) {
+        switch($action) {
             case 'listar':
                 self::listarPedidos();
                 break;
             case 'buscar':
-                self::buscarPedidoPorID();
+                self::buscarPedidoPorId();
                 break;
             case 'cadastrar':
                 self::cadastrarPedido();
@@ -17,44 +16,42 @@ class PedidoController {
             case 'atualizar':
                 self::atualizarPedido();
                 break;
-            case 'excluir':
-                self::excluirPedido();
-                break;
             default:
-                http_response_code(400);
-                echo json_encode(['error' => 'Ação inválida!']);
+                http_response_code(400); // Requisição inválida
+                echo json_encode(['error' => 'Ação inválida']);
                 break;
         }
     }
 
-
-    public static function listarPedidos(){
-        $pedidos= PedidoRepository::getAllPedidos();
+    public static function listarPedidos() {
+        $pedidos = PedidoRepository::getAllPedidos();
         echo json_encode($pedidos);
     }
 
-    public static function buscarPedidoPorID() {
+    public static function buscarPedidoPorId() {
         if($_SERVER['REQUEST_METHOD'] === 'GET') {
             $id = $_GET['id'];
-            $pedido = PedidoRepository::buscarPedidoPorID($id);
+            $pedido = PedidoRepository::getPedidoById($id);
 
             if($pedido) {
-                echo json_encode($pedido);
+
             } else {
                 http_response_code(404);
-                echo json_encode(['error' => 'Produto não econtrado']);
+                echo json_encode(['error' => "Pedido não encontrado!"]);
             }
         } else {
-            http_response_code(405); 
+            http_response_code(405);
+            echo json_encode(['método não permitido' => 'Essa requisição só aceita GET']);
         }
     }
 
     public static function cadastrarPedido() {
         if($_SERVER['REQUEST_METHOD'] === 'POST') {
             $data = json_decode(file_get_contents("php://input"));
-            $pedido = new Pedido(null, $data->nome, $data->descricao, $data->preco);
+            $data_pedido = $data->data_pedido;
+            $status = $data->status;
 
-            $success = PedidoRepository::insertPedido($pedido);
+            $success = PedidoRepository::insertPedido(new Pedido(null, $data_pedido, $status));
             echo json_encode(['success' => $success]);
         } else {
             http_response_code(405);
@@ -64,26 +61,27 @@ class PedidoController {
     public static function atualizarPedido() {
         if($_SERVER['REQUEST_METHOD'] === 'POST') {
             $data = json_decode(file_get_contents("php://input"));
-            $pedido = new Pedido($data->id, $data->nome, $data->descricao, $data->preco);
-
-            $success = PedidoRepository::updatePedido($pedido);
-            echo json_encode(['success' => $success]);
-        } else {
-            http_response_code(405);
-        }
-    }
-
-    public static function excluirPedido() {
-        if($_SERVER['REQUEST_METHOD'] === 'POST') {
-            $data = json_decode(file_get_contents("php://input"));
             $id = $data->id;
-    
-            $success = PedidoRepository::deletePedido($id);
-            echo json_encode(['success' => $success]);
+            $status = $data->status;
+            $data_pedido = $data->data_pedido;
+
+            // Existindo um pedido!
+            $pedidoExistente = PedidoRepository::getPedidoById($id);
+            if($pedidoExistente) {
+                //update das propriedades do pedido
+                $pedidoExistente->setStatus($status);
+                $pedidoExistente->setData($data_pedido);
+
+                $success = PedidoRepository::updatePedido($pedidoExistente, $id);
+                echo json_encode(['success' => $success]);
+            } else {
+                http_response_code(404);
+                echo json_encode(['error' => 'Pedido não encontrado']);
+            }
         } else {
             http_response_code(405);
         }
+        
     }
 }
-
 ?>
